Removed explicit cast to unsigned int because it is implicit in 101-print_number.c

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -19,7 +19,7 @@ void print_number(int n)
 	else
 	{
 		_putchar('-');
-		m = -1 * ((unsigned int) n);
+		m = -1 * n;
 		if (m >= 10)
 			print_number(m / 10);
 		_putchar('0' + m % 10);
